Migrated EH-READ to a separate crate.
