+++ crates/tool-eventhub-read/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "eh-read"
+version = "1.0.0"
+edition = "2021"
+authors = ["Breno RdV"]
+description = "EventHub Message Reader Tool."
+repository = "https://github.com/brenordv/rusted-toolbox"
+
+[dependencies]
+shared = { path = "../shared" }
+clap = "4.5.45"
+anyhow = "1.0.99"
+azeventhubs = "0.20.0"
+chrono = "0.4.41"
+futures-util = "0.3.31"
+sled = "0.34.7"
+tokio = "1.47.1"
+tracing = "0.1.41"
+serde_json = "1.0.142"
+ctrlc = "3.4.7"
+crossterm = "0.27.0"
