+++ crates/tool-eventhub-read/src/eventhub_reader_app.rs
@@ -1,4 +1,3 @@
-use crate::tools::eh_read::progress_tracker::{OperationGuard, ProgressTracker};
 use anyhow::{anyhow, Context, Result};
 use azeventhubs::consumer::{EventPosition, ReadEventOptions};
 use azeventhubs::ReceivedEventData;
@@ -16,6 +15,7 @@ use std::sync::Arc;
 use std::time::Duration;
 use tokio::time::sleep;
 use tracing::{error, info, warn};
+use crate::progress_tracker::{OperationGuard, ProgressTracker};
 
 /// EventHub message consumer with progress tracking and graceful shutdown.
 ///
