+++ crates/tool-eventhub-read/src/graceful_shutdown.rs
@@ -1,8 +1,8 @@
-use crate::tools::eh_read::eventhub_reader_app::EventHubReader;
 use anyhow::Result;
 use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
 use tracing::{error, info};
+use crate::eventhub_reader_app::EventHubReader;
 
 /// Executes graceful shutdown with 15-second timeout.
 ///
