--- crates/tools/src/tools/eh_read/mod.rs
@@ -1,5 +1,0 @@
-pub mod cli_utils;
-pub mod eventhub_reader_app;
-pub mod graceful_shutdown;
-mod progress_tracker;
-pub mod runtime_config_utils;
