+++ crates/tools/src/tools/mod.rs
@@ -1,4 +1,3 @@
-pub mod eh_read;
 pub mod get_lines;
 pub mod guid;
 pub mod jwt;
